Solved some bugs with ordering system and finished promocode system

+++ client/src/utils/payment.ts
@@ -1,3 +1,3 @@
 export const convertToSubcurrency = (num: number) => {
-    return Math.round(parseInt(num.toFixed(2)) * 100);
+    return Math.round(parseFloat(num.toFixed(2)) * 100);
 }
